:bug: Fix several bugs introduced by refactoring
