+++ app/src/commonMain/kotlin/com/crosspaste/ui/devices/DeviceScope.kt
@@ -11,6 +11,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import com.crosspaste.db.sync.SyncRuntimeInfo
 import com.crosspaste.platform.Platform
+import com.crosspaste.sync.SyncHandler
 import com.crosspaste.sync.SyncManager
 import com.crosspaste.ui.base.DialogButtonsView
 import com.crosspaste.ui.base.DialogService
@@ -25,7 +26,10 @@ interface DeviceScope : PlatformScope {
 
     override fun getDeviceDisplayName(): String = syncRuntimeInfo.getDeviceDisplayName()
 
-    fun onEdit(
+    fun getSyncHandler(syncManager: SyncManager): SyncHandler? =
+        syncManager.getSyncHandlers()[syncRuntimeInfo.appInstanceId]
+
+    fun onEditDevice(
         dialogService: DialogService,
         pasteDialogFactory: PasteDialogFactory,
         syncManager: SyncManager,
@@ -46,7 +50,7 @@ interface DeviceScope : PlatformScope {
                     if (inputNoteName == "") {
                         isError = true
                     } else {
-                        syncManager.getSyncHandlers()[syncRuntimeInfo.appInstanceId]?.let { syncHandler ->
+                        getSyncHandler(syncManager)?.let { syncHandler ->
                             syncHandler.updateNoteName(inputNoteName) {}
                         }
                         dialogService.popDialog()
