+++ app/src/commonMain/kotlin/com/crosspaste/ui/devices/PlatformScope.kt
@@ -25,5 +25,5 @@ interface PlatformScope {
     fun Modifier.hoverModifier(
         onEnter: () -> Unit = {},
         onExit: () -> Unit = {},
-    ): Modifier
+    ): Modifier = this
 }
